+++ README.md
@@ -1,5 +1,5 @@
 # This is the repository of the system for the Thesis of the Group Computer Science Global Offensive
-## This Branch is currently in Version 1.0.3
+## This Branch is currently in Version 1.0.4
 ### This branch focuses on building of the GUI 
 
 Members of the thesis group are:
