+++ views/README.md
@@ -1,0 +1,7 @@
+# Structure of the GUI
+For Reference click the [GUI flow](https://drive.google.com/file/d/1-EQPn6qUzVsC2B4StnYKVexv7ps9LH9F/view?usp=sharing)
+
+## Branch Current Version is 1.0.4
+
+To do:
+    Integrate the functions of control to Views.
